+++ trecord/src/apis/Comment/getReplyComment.ts
@@ -1,6 +1,6 @@
 import { UseQueryResult, useQuery } from '@tanstack/react-query';
 import { http } from '../_http';
-import { GetReplyCommentProps } from '@/types/comment';
+import { GetReplyCommentProps, GetReplyCommentResponse } from '@/types/comment';
 import COMMENT_API_KEY from './constants';
 
 interface Props {
@@ -8,15 +8,15 @@ interface Props {
 }
 export const getReplyComment = async ({
   commentId,
-}: Props): Promise<GetReplyCommentProps[]> => {
+}: Props): Promise<GetReplyCommentResponse> => {
   const url = `/api/v1/comments/${commentId}/replies`;
-  const response: { content: GetReplyCommentProps[] } = await http.get(url);
-  return response.content;
+  const response: GetReplyCommentResponse = await http.get(url);
+  return response;
 };
 
 const useGetReplyComment = ({
   commentId,
-}: Props): UseQueryResult<GetReplyCommentProps[]> => {
+}: Props): UseQueryResult<GetReplyCommentResponse> => {
   return useQuery(
     [COMMENT_API_KEY.REPLY_COMMENT, { comment_id: commentId }],
     () => getReplyComment({ commentId }),
